+++ api/src/pdc/services/apdf/actions/ExportAction.ts
@@ -5,9 +5,9 @@ import { logger } from "@/lib/logger/index.ts";
 import { get } from "@/lib/object/index.ts";
 import { internalOnlyMiddlewares } from "@/pdc/providers/middleware/index.ts";
 import { BucketName, S3StorageProvider } from "@/pdc/providers/storage/index.ts";
-import { handlerConfig, ParamsInterface, ResultInterface } from "@/shared/apdf/export.contract.ts";
-import { alias } from "@/shared/apdf/export.schema.ts";
 import { ResultInterface as PolicyResultInterface } from "../../policy/contracts/find.contract.ts";
+import { handlerConfig, ParamsInterface, ResultInterface } from "../contracts/export.contract.ts";
+import { alias } from "../contracts/export.schema.ts";
 import { castExportParams } from "../helpers/castExportParams.helper.ts";
 import { getCampaignOperators } from "../helpers/getCampaignOperators.helper.ts";
 import { DataRepositoryProviderInterfaceResolver } from "../interfaces/APDFRepositoryProviderInterface.ts";
