+++ api/src/pdc/services/apdf/providers/APDFRepositoryProvider.ts
@@ -3,9 +3,9 @@ import { provider } from "@/ilos/common/index.ts";
 import { PostgresConnection } from "@/ilos/connection-postgres/index.ts";
 import { set } from "@/lib/object/index.ts";
 import sql, { raw } from "@/lib/pg/sql.ts";
-import { PolicyStatsInterface } from "@/shared/apdf/interfaces/PolicySliceStatInterface.ts";
 import { PgCursorHandler } from "@/shared/common/PromisifiedPgCursor.ts";
 import { UnboundedSlices } from "../../policy/contracts/common/interfaces/Slices.ts";
+import { PolicyStatsInterface } from "../contracts/interfaces/PolicySliceStatInterface.ts";
 import {
   CampaignSearchParamsInterface,
   DataRepositoryInterface,
