+++ api/src/pdc/services/apdf/providers/excel/BuildExcel.ts
@@ -2,8 +2,8 @@ import { excel } from "@/deps.ts";
 import { KernelInterfaceResolver, provider } from "@/ilos/common/index.ts";
 import { logger } from "@/lib/logger/index.ts";
 import { APDFNameProvider } from "@/pdc/providers/storage/index.ts";
-import { SliceStatInterface } from "@/shared/apdf/interfaces/PolicySliceStatInterface.ts";
 import { ResultInterface as Campaign } from "../../../policy/contracts/find.contract.ts";
+import { SliceStatInterface } from "../../contracts/interfaces/PolicySliceStatInterface.ts";
 import { CampaignSearchParamsInterface } from "../../interfaces/APDFRepositoryProviderInterface.ts";
 import { DataRepositoryProvider } from "../APDFRepositoryProvider.ts";
 import { SlicesWorksheetWriter } from "./SlicesWorksheetWriter.ts";
